update Go-style comments and change by suggestion
